[common/meta/raft-store] feature: impl MetaAPI for StateMachine
StateMachine already provided meta-api methods thus it is natural to
imopl MetaApi for it.

And this let other upper level mods be able to reuse it.
E.g., flight action handler do not need to re-impl MetaApi.

Now MetaEmbedded is purely a wrapper of StateMachine
